move wrapText
