--- markdown/h2m/utils/wrap.js
@@ -1,5 +1,0 @@
-function wrapText(h, value) {
-  return h.wrapText ? value : value.replace(/\r?\n|\r/g, " ");
-}
-
-module.exports = wrapText;
