+++ package.json
@@ -96,7 +96,6 @@
     "send": "0.17.1",
     "source-map-support": "0.5.19",
     "tempy": "1.0.1",
-    "trim-trailing-lines": "^2.0.0",
     "unified": "^9.2.1",
     "unist-builder": "^2.0.3",
     "unist-util-visit": "2.0.3"
